Made changes to the way logging and reporting work
o	Log now has test step number
o	CSV is saved at the same time as the report for errors instead of after running the test step. This is so that re ran failed test cases that pass are recorded as pass.
o	The time for the reporter has been updated